Set up a NavController, created a destination file, created a NavGraph, and a simple Account Page
Navigation is the way of connecting buttons to different pages, and this commit is a set up of the navigation function. This commit also created a simple Account Page which is one of the four main pages designed for Sweetea App.

Futher problems: Homescreen page has not separated from the app and added to the NavGraph yet

Resolved: #29 & #16